Move global-dir out of core
