+++ v-next/hardhat/src/internal/global-dir.ts
